+++ sickbeard/providers/ettv.py
@@ -15,10 +15,8 @@
 # You should have received a copy of the GNU General Public License
 # along with SickGear.  If not, see <http://www.gnu.org/licenses/>.
 
-import base64
 import re
 import traceback
-import urllib
 
 from . import generic
 from sickbeard import logger
@@ -27,29 +25,23 @@ from sickbeard.helpers import tryInt
 from lib.unidecode import unidecode
 
 
-class BTSceneProvider(generic.TorrentProvider):
+class ETTVProvider(generic.TorrentProvider):
 
     def __init__(self):
-        generic.TorrentProvider.__init__(self, 'BTScene')
+        generic.TorrentProvider.__init__(self, 'ETTV')
 
-        self.url_home = ['http://btsone.cc/',  'http://diriri.xyz/'] + \
-                        ['https://%s/' % base64.b64decode(x) for x in [''.join(x) for x in [
-                            [re.sub('[L\sT]+', '', x[::-1]) for x in [
-                                'zTRnTY', 'uVT 2Y', '15LSTZ', 's JmLb', 'rTNL2b', 'uQW LZ', '=LLMmd']],
-                            [re.sub('[j\sq]+', '', x[::-1]) for x in [
-                                'zRn qY', 'l52j b', '1j5S M', 'sq Jmb', 'r Nq2b', 'ujQWqZ', 's9jGqb']],
-                        ]]]
-        self.url_vars = {'search': '?q=%s&category=series&order=1', 'browse': 'lastdaycat/type/Series/',
-                         'get': 'torrentdownload.php?id=%s'}
-        self.url_tmpl = {'config_provider_home_uri': '%(home)s', 'search': '%(vars)s',
-                         'browse': '%(home)s%(vars)s', 'get': '%(home)s%(vars)s'}
+        self.url_home = ['https://ettv.tv/']
+        self.url_vars = {'search': 'torrents-search.php?%s&search=%s&sort=id&order=desc'}
+        self.url_tmpl = {'config_provider_home_uri': '%(home)s', 'search': '%(home)s%(vars)s'}
+
+        self.categories = {'Season': [7], 'Episode': [41, 5, 50]}
+        self.categories['Cache'] = self.categories['Season'] + self.categories['Episode']
 
         self.minseed, self.minleech = 2 * [None]
-        self.confirmed = False
 
     @staticmethod
     def _has_signature(data=None):
-        return data and re.search(r'(?i)(?:btscene|bts[-]official|full\sindex)', data)
+        return data and re.search(r'(?i)(?:ettv)', data)
 
     def _search_provider(self, search_params, **kwargs):
 
@@ -59,26 +51,15 @@ class BTSceneProvider(generic.TorrentProvider):
 
         items = {'Cache': [], 'Season': [], 'Episode': [], 'Propers': []}
 
-        rc = dict((k, re.compile('(?i)' + v)) for (k, v) in {
-            'info': '\w+?(\d+)[.]html', 'verified': 'Verified'}.iteritems())
-
-        url = self.url
-        response = self.get_url(url)
-        form = re.findall('(?is)(<form[^>]+)', response)
-        response = any(form) and form[0] or response
-        action = re.findall('<form[^>]+action=[\'"]([^\'"]*)', response)[0]
-        url = action if action.startswith('http') else \
-            url if not action else \
-            (url + action) if action.startswith('?') else \
-            self.urls['config_provider_home_uri'] + action.lstrip('/')
+        rc = dict((k, re.compile('(?i)' + v)) for (k, v) in {'info': 'torrent/'}.iteritems())
 
         for mode in search_params.keys():
             for search_string in search_params[mode]:
 
                 search_string = isinstance(search_string, unicode) and unidecode(search_string) or search_string
 
-                search_url = self.urls['browse'] if 'Cache' == mode \
-                    else url + self.urls['search'] % (urllib.quote_plus(search_string))
+                search_url = self.urls['search'] % (
+                    self._categories_string(mode), ('%2B ', '')['Cache' == mode] + '.'.join(search_string.split()))
 
                 html = self.get_url(search_url)
 
@@ -87,33 +68,28 @@ class BTSceneProvider(generic.TorrentProvider):
                     if not html or self._has_no_results(html):
                         raise generic.HaltParseException
                     with BS4Parser(html, features=['html5lib', 'permissive']) as soup:
-                        torrent_rows = soup.select('tr[class$="_tr"]')
+                        torrent_table = soup.find('table', class_='table')
+                        torrent_rows = [] if not torrent_table else torrent_table.find_all('tr')
 
                         if not len(torrent_rows):
                             raise generic.HaltParseException
 
                         head = None
-                        for tr in torrent_rows:
+                        for tr in torrent_rows[1:]:
                             cells = tr.find_all('td')
                             if 6 > len(cells):
                                 continue
                             try:
-                                head = head if None is not head else self._header_row(tr)
+                                head = head if None is not head else self._header_row(
+                                    tr, {'seed': r'seed', 'leech': r'leech', 'size': r'^size'})
                                 seeders, leechers, size = [tryInt(n, n) for n in [
                                     cells[head[x]].get_text().strip() for x in 'seed', 'leech', 'size']]
-                                if self._peers_fail(mode, seeders, leechers) or \
-                                        self.confirmed and not (tr.find('img', src=rc['verified'])
-                                                                or tr.find('img', title=rc['verified'])):
+                                if self._peers_fail(mode, seeders, leechers):
                                     continue
 
                                 info = tr.find('a', href=rc['info'])
-                                title = info and info.get_text().strip()
-                                tid_href = info and rc['info'].findall(info['href'])
-                                tid_href = tid_href and tryInt(tid_href[0], 0) or 0
-                                tid_tr = tryInt(tr['id'].strip('_'), 0)
-                                tid = (tid_tr, tid_href)[tid_href > tid_tr]
-
-                                download_url = info and (self.urls['get'] % tid)
+                                title = (info.attrs.get('title') or info.get_text()).strip()
+                                download_url = self._link(info.get('href'))
                             except (AttributeError, TypeError, ValueError, IndexError):
                                 continue
 
@@ -131,8 +107,23 @@ class BTSceneProvider(generic.TorrentProvider):
 
         return results
 
-    def _episode_strings(self, ep_obj, **kwargs):
-        return super(BTSceneProvider, self)._episode_strings(ep_obj, sep_date='.', **kwargs)
+    def get_data(self, url):
+        result = None
+        html = self.get_url(url, timeout=90)
+        try:
+            result = re.findall('(?i)"(magnet:[^"]+?)">', html)[0]
+        except IndexError:
+            logger.log('Failed no magnet in response', logger.DEBUG)
+        return result
+
+    def get_result(self, episodes, url):
+        result = None
+
+        if url:
+            result = super(ETTVProvider, self).get_result(episodes, url)
+            result.get_data_func = self.get_data
+
+        return result
 
 
-provider = BTSceneProvider()
+provider = ETTVProvider()